Remove unneeded files from dpdk-inline

+++ os/src/net/dpdk-inline/rte_os.h
@@ -1,9 +1,7 @@
 #pragma once
 
-// This header contains a polyfill for our stub OS, i.e., stuff not otherwise defined but that DPDK expects of all OSes
-
 #include <stdint.h>
 
 
-// We don't support multicore for now so this can be anything
-typedef uint64_t rte_cpuset_t;
+// OS ASSUMPTION: Single core -> this can be anything
+typedef uintptr_t rte_cpuset_t;
